seperating the form from the modal
